requirements.txt
fastapi
uvicorn
pymysql
boto3
psycopg2-binary
pyodbc
cx_Oracle
ibm_db

